Creada herencia Persona, para alumnos y profesores
