+++ src/com/IesAlonsoMadrigal/domain/models/Alumnos.java
@@ -1,55 +1,9 @@
 package com.IesAlonsoMadrigal.domain.models;
 
-public class Alumnos {
-
-    private Integer id;
-    private String nombre;
-    private String apellidos;
-    private String dni;
-    private String direccion;
+public class Alumnos extends Persona{
     private String madre;
     private String padre;
     private String email;
-
-    public Integer getId() {
-        return id;
-    }
-
-    public void setId(Integer id) {
-        this.id = id;
-    }
-
-    public String getNombre() {
-        return nombre;
-    }
-
-    public void setNombre(String nombre) {
-        this.nombre = nombre;
-    }
-
-    public String getApellidos() {
-        return apellidos;
-    }
-
-    public void setApellidos(String apellidos) {
-        this.apellidos = apellidos;
-    }
-
-    public String getDni() {
-        return dni;
-    }
-
-    public void setDni(String dni) {
-        this.dni = dni;
-    }
-
-    public String getDireccion() {
-        return direccion;
-    }
-
-    public void setDireccion(String direccion) {
-        this.direccion = direccion;
-    }
 
     public String getMadre() {
         return madre;
