+++ src/com/IesAlonsoMadrigal/domain/models/Profesores.java
@@ -2,26 +2,9 @@ package com.IesAlonsoMadrigal.domain.models;
 
 public class Profesores {
 
-    private Integer id;
-    private String nombre;
     private String titulo_academico;
     private String fecha_alta;
 
-    public Integer getId() {
-        return id;
-    }
-
-    public void setId(Integer id) {
-        this.id = id;
-    }
-
-    public String getNombre() {
-        return nombre;
-    }
-
-    public void setNombre(String nombre) {
-        this.nombre = nombre;
-    }
 
     public String getTitulo_academico() {
         return titulo_academico;
